move cmd to cmd/kubeapps
