+++ cmd/kubeapps/dashboard_test.go
@@ -1,4 +1,4 @@
-package cmd
+package kubeapps
 
 import (
 	"testing"
