+++ cmd/kubeapps/up_test.go
@@ -1,4 +1,4 @@
-package cmd
+package kubeapps
 
 import (
 	"bytes"
